Added some ToT portal scripts. Fixed not getting sp past level 127. "Fixed" auto completing quests that shouldn't be autocompleted. Added all clock packets.

+++ src/main/java/net/swordie/ms/client/character/Char.java
@@ -1475,7 +1475,7 @@ public class Char {
 	public void addSpToJobByCurrentLevel(int num) {
 		CharacterStat cs = getAvatarData().getCharacterStat();
 		if (JobConstants.isExtendSpJob(getJob())) {
-			byte jobLevel = (byte) JobConstants.getJobLevelByCharLevel((byte) getLevel());
+			byte jobLevel = (byte) JobConstants.getJobLevelByCharLevel(getLevel());
 			num += cs.getExtendSP().getSpByJobLevel(jobLevel);
 			getAvatarData().getCharacterStat().getExtendSP().setSpToJobLevel(jobLevel, num);
 		} else {
